Added: Delete file/folder

+++ LeoboxV2/node.cs
@@ -12,7 +12,7 @@ namespace LeoboxV2
         public string mime_type { get; set; }
         public string name { get; set; }
         public string path_file { get; set; }
-        public int size { get; set; }
+        public Int64 size { get; set; }
         public Int64 storage_mtime { get; set; }
         public List<node> sub_dir { get; set; }
         public string type { get; set; }
